+++ mne/__init__.py
@@ -63,6 +63,7 @@ from .source_space import (read_source_spaces, vertex_to_mni,
                            setup_volume_source_space, SourceSpaces,
                            add_source_space_distances, morph_source_spaces,
                            get_volume_labels_from_aseg)
+from .annotations import Annotations
 from .epochs import Epochs, EpochsArray, read_epochs
 from .evoked import Evoked, EvokedArray, read_evokeds, write_evokeds, combine_evoked
 from .label import (read_label, label_sign_flip,